TextField,Buttonの設定をextensionに移動
