+++ 書籍管理アプリ-iOS/BooksViewController.swift
@@ -1,7 +1,7 @@
 import UIKit
 
 class BooksViewController: UIViewController, UITableViewDelegate, UITableViewDataSource {
-
+    
     var books:[Book] = []
     
     private lazy var bookTableView: UITableView = {
@@ -15,23 +15,19 @@ class BooksViewController: UIViewController, UITableViewDelegate, UITableViewDat
     
     private lazy var loadButton: UIButton = {
         let button = UIButton()
-        let loadTitle = R.string.localizable.load()
-        button.layer.position = CGPoint(x:self.view.frame.width/2, y:200)
-        button.setTitle(loadTitle, for: .normal)
-        button.setTitleColor( .lightGray, for: .normal)
-        button.titleLabel?.font =  .systemFont(ofSize: 24)
-        button.backgroundColor = UIColor(red:0.9, green: 0.9, blue: 0.9, alpha: 1)
+        button.setTitle(R.string.localizable.load(), for: .normal)
+        button.setButton()
         return button
     }()
     
     override func viewDidLoad() {
         super.viewDidLoad()
-
+        
         title = R.string.localizable.booksview()
         
         view.addSubview(bookTableView)
         view.addSubview(loadButton)
-
+        
         fetchData()
         
         //追加ボタンの設定
@@ -48,7 +44,7 @@ class BooksViewController: UIViewController, UITableViewDelegate, UITableViewDat
         loadButton.widthAnchor.constraint(equalTo:self.view.widthAnchor).isActive = true
         
     }
-
+    
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
     }
@@ -67,29 +63,29 @@ class BooksViewController: UIViewController, UITableViewDelegate, UITableViewDat
         addBookViewController.modalTransitionStyle = .crossDissolve
         present(navi, animated: true, completion: nil)
     }
-
+    
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-
+        
         let cell = tableView.dequeueReusableCell(withIdentifier: NSStringFromClass(BookCell.self), for: indexPath) as! BookCell
         
         cell.bookRegister(book: books[indexPath.row])
-
+        
         return cell
     }
-
+    
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return 4
     }
-
+    
     //書籍編集画面への遷移処理
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-
+        
         let editBookViewController = EditBookViewController()
         
         editBookViewController.book = books[indexPath.row]
         
         self.navigationController?.pushViewController(editBookViewController, animated: true)
-
+        
     }
-
+    
 }
